+++ lib/vtls/vtls.h
@@ -36,6 +36,8 @@ struct Curl_ssl {
 
   unsigned support_https_proxy:1; /* supports access via HTTPS proxies */
 
+  size_t sizeof_ssl_backend_data;
+
   int (*init)(void);
   void (*cleanup)(void);
